Update community name in a couple places
There are other places in the document that refer to the authority
figures in a community (e.g. "the [COMMUNITY] organizers").

For events, that's the event organizers. I'm not sure who the authority
in the Open Humans community is, whether it's Open Humans Foundation
staff or community moderators, or someone else.