Add open issues document and power structure questions.
When thinking about who will enforce a Code of Conduct, it's important
to think in terms of the power structure a community is putting in
place. Many communities mix the power structure that decides technical
or financial direction with the power structure that enforces the Code
of Conduct. This often leads to conflicts of interest when evaluating
incident reports and following up with reported people.