Add familial status and pregnancy as protected classes.
Community members of Open Humans may want to upload data related to
reproduction, such as basal body temperature to measure when ovulation
happens. It's important not to harass someone for their choices around
reproduction, family, or even choosing to not have children.

These protected classes were found in the Carpentries Code of Conduct.