Add citizenship and immigration status as protected classes.
Community members may upload location data to Open Humans, which could
potentially be used to harass citizens of a specific country, or used to
harass someone by reporting them to immigration enforcement. It's
important to call out that harassing someone on the basis of their
citizenship, nationality, or immigration status is not appropriate.

The citizenship protected class was found in the Data Carpentries Code
of Conduct, and the immigration status protected class was found in the
LGBTQ in Tech Slack Code of Conduct.